requirements.txt
Flask
langchain
langchain-community
langchain-groq
python-dotenv
PyMuPDF
